Renamed fall_detection.py to webcam_demo.py
